# controller/response.go
package controller

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

/*

{
	"code": 200,  	// 程序中的错误码
	"msg": "xxxx"   // 提示信息
	"data": {}      // 数据
}

*/

type ResponseData struct {
	Code ResCode		`json:"code"`
	Msg  interface{}	`json:"msg"`
	Data interface{}	`json:"data,omitempty"`     // omitempty 忽略为空的值 为空时在json中不显示
}

func ResponseError(c *gin.Context, code ResCode) {
	rd := &ResponseData{
		Code: code,
		Msg:  code.Msg(),
		Data: nil,
	}
	c.JSON(http.StatusOK, rd)
}

func ResponseErrorWithMsg(c *gin.Context, code ResCode, msg interface{}) {
	rd := &ResponseData{
		Code: code,
		Msg:  msg,
		Data: nil,
	}
	c.JSON(http.StatusOK, rd)
}

func ResponseSuccess(c *gin.Context, data interface{}) {
	rd := &ResponseData{
		Code: CodeSuccess,
		Msg:  CodeSuccess.Msg(),
		Data: data,
	}
	c.JSON(http.StatusOK, rd)
}